Refactored SerializationHelper
Moved Base related methods to BaseSerializationHelper
Fixed overflow bug in CompressStream() and added tests for verification
Removed unused code in SerializationHelper
